+++ src/directives/tableGridBodyCellTemplate/tableGridBodyCellTemplate.directive.ts
@@ -1,6 +1,6 @@
-import {Directive} from '@angular/core';
+import {Directive, TemplateRef} from '@angular/core';
 
-import {TableGridCellTemplateBaseDirective} from '../tableGridCellTemplate/tableGridCellTemplate.directive';
+import {DataCellTemplateContext} from '../../interfaces';
 
 /**
  * Directive that is used for obtaining template for table grid body cell
@@ -10,6 +10,21 @@ import {TableGridCellTemplateBaseDirective} from '../tableGridCellTemplate/table
     selector: '[bodyCellTemplate]',
     standalone: true,
 })
-export class TableGridBodyCellTemplateSADirective<TData = unknown> extends TableGridCellTemplateBaseDirective<TData>
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export class TableGridBodyCellTemplateSADirective<TColumnMetadata = any, TData = any>
 {
+    //######################### constructor #########################
+    constructor(public template: TemplateRef<DataCellTemplateContext<TColumnMetadata, TData>>)
+    {
+    }
+
+    //######################### ng language server #########################
+    
+    /**
+     * Allows typechecking for template
+     */
+    static ngTemplateContextGuard(_dir: TableGridBodyCellTemplateSADirective, _ctx: unknown): _ctx is DataCellTemplateContext
+    {
+        return true;
+    }
 }