+++ src/directives/tableGridHeaderCellTemplate/tableGridHeaderCellTemplate.directive.ts
@@ -1,6 +1,7 @@
-import {Directive} from '@angular/core';
+import {Directive, TemplateRef} from '@angular/core';
 
-import {TableGridCellTemplateBaseDirective} from '../tableGridCellTemplate/tableGridCellTemplate.directive';
+import {CellTemplateContext} from '../../interfaces';
+
 
 /**
  * Directive that is used for obtaining template for table grid header cell
@@ -10,6 +11,21 @@ import {TableGridCellTemplateBaseDirective} from '../tableGridCellTemplate/table
     selector: '[headerCellTemplate]',
     standalone: true,
 })
-export class TableGridHeaderCellTemplateSADirective<TData = unknown> extends TableGridCellTemplateBaseDirective<TData>
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export class TableGridHeaderCellTemplateSADirective<TColumnMetadata = any>
 {
+    //######################### constructor #########################
+    constructor(public template: TemplateRef<CellTemplateContext<TColumnMetadata>>)
+    {
+    }
+
+    //######################### ng language server #########################
+    
+    /**
+     * Allows typechecking for template
+     */
+    static ngTemplateContextGuard(_dir: TableGridHeaderCellTemplateSADirective, _ctx: unknown): _ctx is CellTemplateContext
+    {
+        return true;
+    }
 }