+++ src/interfaces/cellTemplateContext/cellTemplate.context.ts
@@ -1,0 +1,13 @@
+/**
+ * Context for cell template
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export interface CellTemplateContext<TColumnMetadata = any>
+{
+    //######################### readonly properties #########################
+
+    /**
+     * Object of column metadata itself
+     */
+    readonly column: TColumnMetadata;
+}