+++ src/pipes/index.ts
@@ -1,2 +1,4 @@
+export * from './cellContext/cellContext.pipe';
+export * from './dataCellContext/dataCellContext.pipe';
 export * from './readValue/readValue.pipe';
 export * from './resolveForwardRef/resolveForwardRef.pipe';